backend/requirements-fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.24.0
motor==3.3.1
pydantic==2.5.0
pydantic-settings==2.0.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
python-decouple==3.8
Pillow==10.1.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
jdatetime==4.1.0